Add `Link`s constraints
It's always "efficiency * input == output" but since the `Link` is
bidirectional, we have to differentiate between what constitutes the
input and what constitutes the output, which is kinda difficult within
the scope of our current terminology. I tried doing it using the sign of
the flow, But I think this is wrong.