Wrap lines to be below 80 characters
I also took the liberty of removing a pair of curly braces wrapped
around the complete lite. These braces don't actually do anything, so
removing them shouldn't have a noticeable effect. And I removed the
space in the `\frac{ }{}` construct which used in the last commit in
order to make the two versions easier to compare.