Append `\forall` line instead of prepending it
This streamlines urbs' storage formulation with how the other frameworks
are writing it. I also took the liberty of removing the period at the
end of the equation, since an equation is not a sentence.
Sorry for being invasive. Feel free to revert if this is not to your
liking. :)