Use only one `\forall` line for all constraints
Sorry for being so invasive again, but since the lines are pretty much
equal, it's IMHO more readable to only have one line covering all
constraints. This also reduces the chance for typos, of which there was
on in the last line. Just as a heads up in case anyone decides to revert
this, which you're obviously free to do. :)